requirements.txt
fastapi
google-cloud-storage
python-dotenv
google-generativeai
python-multipart
PyPDF2